requirements.txt
Django==6.0.1
djangorestframework==3.16.1
cryptography==46.0.3
PyJWT==2.10.1
better-profanity==0.7.0
pillow==12.1.0
python-dotenv==1.2.1
dj-database-url==3.1.0
psycopg2-binary==2.9.11
redis==7.1.0
channels_redis==4.3.0
uvicorn==0.40.0
websockets==16.0
gunicorn==24.1.1
channels==4.3.2
daphne==4.2.1
whitenoise==6.11.0
tabulate==0.9.0
spacy==3.8.11
honcho==2.0.0